# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoCopyAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
  /billing/topup:
    post:
      operationId: topupCredits
      tags: [billing]
      summary: Purchase a one-time credit top-up pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupRequest"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/users/{id}/credits:
    patch:
      operationId: adminAdjustCredits
      tags: [admin]
      summary: Adjust user credits (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditAdjustment"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/role:
    patch:
      operationId: adminUpdateRole
      tags: [admin]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/generations:
    get:
      operationId: adminListGenerations
      tags: [admin]
      summary: List all AI generations (admin only)
      parameters:
        - name: user_id
          in: query
          schema:
            type: integer
        - name: tool_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Generations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiGeneration"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, email, role, credits, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        businessName:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        credits:
          type: number
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        currentPlan:
          type: ["string", "null"]
        planStatus:
          type: ["string", "null"]
        createdAt:
          type: string
        lastLoginAt:
          type: ["string", "null"]
...
    AiGeneration:
      type: object
      required: [id, toolType, inputSummary, output, creditsUsed, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        toolType:
          type: string
        inputSummary:
          type: string
        output:
          type: string
        tokensEstimate:
          type: ["integer", "null"]
        creditsUsed:
          type: number
        createdAt:
          type: string
...
    TopupRequest:
      type: object
      required: [packId]
      properties:
        packId:
          type: string
    CheckoutRequest:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
    CheckoutResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    AdminUsersResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
    CreditAdjustment:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        reason:
          type: string
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [user, admin]
...